+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## develop branch
 
-Nothing yet.
+### New
+
+* Added `Checks\IsGitRepo`